requirements.txt
# Core data manipulation and analysis
pandas>=1.3.0
numpy>=1.21.0

# Machine learning and metrics
scikit-learn>=1.0.0
joblib>=1.0.0

# Natural language processing
nltk>=3.6.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Sparse matrix operations
scipy>=1.7.0

# Hugging Face integration
huggingface-hub>=0.8.0

# Environment variables
python-dotenv>=0.19.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
tqdm>=4.60.0
requests>=2.25.0
certifi>=2021.5.30
urllib3>=1.26.0

# Testing
pytest>=7.0.0